Add host service para chamada de api da OpenWeather

+++ DataStorage/Model/Weather.cs
@@ -4,6 +4,6 @@
     {
         public string Cidade { get; set; } = string.Empty;
         public DateTime Data { get; set; }
-        public float Temperatura { get; set; }
+        public double Temperatura { get; set; }
     }
 }
