+++ Domain/Query/GetWeatherListQueryResult.cs
@@ -18,6 +18,6 @@ namespace Domain.Query
 
         public string Cidade { get; set; } = string.Empty;
         public DateTime Data { get; set; }
-        public float Temperatura { get; set; }
+        public double Temperatura { get; set; }
     }
 }
